+++ assets/js/dashboard/stats/pages/pages.js
@@ -29,6 +29,10 @@ export default class Visits extends React.Component {
     }
   }
 
+  showConversionRate() {
+    return !!this.props.query.filters.goal
+  }
+
   fetchPages() {
     api.get(`/api/stats/${encodeURIComponent(this.props.site.domain)}/pages`, this.props.query)
       .then((res) => this.setState({loading: false, pages: res}))
@@ -39,6 +43,7 @@ export default class Visits extends React.Component {
     query.set('page', page.name)
     const domain = new URL('https://' + this.props.site.domain)
     const externalLink = 'https://' + domain.host  + page.name
+    const maxWidthDeduction =  this.showConversionRate() ? "10rem" : "5rem"
 
     return (
       <div
@@ -49,7 +54,7 @@ export default class Visits extends React.Component {
           count={page.count}
           all={this.state.pages}
           bg="bg-orange-50 dark:bg-gray-500 dark:bg-opacity-15"
-          maxWidthDeduction="4rem"
+          maxWidthDeduction={maxWidthDeduction}
         >
           <span
             className="flex px-2 py-1.5 group dark:text-gray-300 relative z-9 break-all"
@@ -69,7 +74,8 @@ export default class Visits extends React.Component {
             </a>
           </span>
         </Bar>
-        <span className="font-medium dark:text-gray-200">{numberFormatter(page.count)}</span>
+        <span className="font-medium dark:text-gray-200 w-20 text-right">{numberFormatter(page.count)}</span>
+        {this.showConversionRate() && <span className="font-medium dark:text-gray-200 w-20 text-right">{numberFormatter(page.conversion_rate)}%</span>}
       </div>
     )
   }
@@ -89,7 +95,10 @@ export default class Visits extends React.Component {
         <React.Fragment>
           <div className="flex items-center justify-between mt-3 mb-2 text-xs font-bold tracking-wide text-gray-500 dark:text-gray-400">
             <span>Page url</span>
-            <span>{ this.label() }</span>
+            <div className="text-right">
+              <span className="inline-block w-20">{ this.label() }</span>
+              {this.showConversionRate() && <span className="inline-block w-20">CR</span>}
+            </div>
           </div>
 
           <FlipMove>