+++ assets/js/dashboard/stats/sources/source-list.js
@@ -33,14 +33,20 @@ class AllSources extends React.Component {
     return this.props.query.period === 'realtime'
   }
 
+  showConversionRate() {
+    return !!this.props.query.filters.goal
+  }
+
   fetchReferrers() {
     api.get(`/api/stats/${encodeURIComponent(this.props.site.domain)}/sources`, this.props.query, {show_noref: this.showNoRef()})
-      .then((res) => this.setState({loading: false, referrers: res}))
+       .then((res) => this.setState({loading: false, referrers: res}))
   }
 
   renderReferrer(referrer) {
     const query = new URLSearchParams(window.location.search)
     query.set('source', referrer.name)
+
+    const maxWidthDeduction =  this.showConversionRate() ? "10rem" : "5rem"
 
     return (
       <div
@@ -51,7 +57,7 @@ class AllSources extends React.Component {
           count={referrer.count}
           all={this.state.referrers}
           bg="bg-blue-50 dark:bg-gray-500 dark:bg-opacity-15"
-          maxWidthDeduction="4rem"
+          maxWidthDeduction={maxWidthDeduction}
         >
           <span className="flex px-2 py-1.5 dark:text-gray-300 relative z-9 break-all">
             <Link
@@ -66,7 +72,8 @@ class AllSources extends React.Component {
             </Link>
           </span>
         </Bar>
-        <span className="font-medium dark:text-gray-200">{numberFormatter(referrer.count)}</span>
+        <span className="font-medium dark:text-gray-200 w-20 text-right">{numberFormatter(referrer.count)}</span>
+        {this.showConversionRate() && <span className="font-medium dark:text-gray-200 w-20 text-right">{referrer.conversion_rate}%</span>}
       </div>
     )
   }
@@ -81,7 +88,10 @@ class AllSources extends React.Component {
         <React.Fragment>
           <div className="flex items-center justify-between mt-3 mb-2 text-xs font-bold tracking-wide text-gray-500">
             <span>Source</span>
-            <span>{this.label()}</span>
+            <div className="text-right">
+              <span className="inline-block w-20">{this.label()}</span>
+              {this.showConversionRate() && <span className="inline-block w-20">CR</span>}
+            </div>
           </div>
 
           <FlipMove className="flex-grow">
@@ -149,6 +159,10 @@ class UTMSources extends React.Component {
     return this.props.query.period === 'realtime'
   }
 
+  showConversionRate() {
+    return !!this.props.query.filters.goal
+  }
+
   fetchReferrers() {
     const endpoint = UTM_TAGS[this.props.tab].endpoint
     api.get(`/api/stats/${encodeURIComponent(this.props.site.domain)}/${endpoint}`, this.props.query, {show_noref: this.showNoRef()})
@@ -158,6 +172,7 @@ class UTMSources extends React.Component {
   renderReferrer(referrer) {
     const query = new URLSearchParams(window.location.search)
     query.set(this.props.tab, referrer.name)
+    const maxWidthDeduction =  this.showConversionRate() ? "10rem" : "5rem"
 
     return (
       <div
@@ -168,7 +183,7 @@ class UTMSources extends React.Component {
           count={referrer.count}
           all={this.state.referrers}
           bg="bg-blue-50 dark:bg-gray-500 dark:bg-opacity-15"
-          maxWidthDeduction="4rem"
+          maxWidthDeduction={maxWidthDeduction}
         >
 
           <span className="flex px-2 py-1.5 dark:text-gray-300 relative z-9 break-all">
@@ -180,7 +195,8 @@ class UTMSources extends React.Component {
             </Link>
           </span>
         </Bar>
-        <span className="font-medium dark:text-gray-200">{numberFormatter(referrer.count)}</span>
+        <span className="font-medium dark:text-gray-200 w-20 text-right">{numberFormatter(referrer.count)}</span>
+        {this.showConversionRate() && <span className="font-medium dark:text-gray-200 w-20 text-right">{referrer.conversion_rate}%</span>}
       </div>
     )
   }
@@ -195,7 +211,10 @@ class UTMSources extends React.Component {
         <div className="flex flex-col flex-grow">
           <div className="flex items-center justify-between mt-3 mb-2 text-xs font-bold tracking-wide text-gray-500 dark:text-gray-400">
             <span>{UTM_TAGS[this.props.tab].label}</span>
-            <span>{this.label()}</span>
+            <div className="text-right">
+              <span className="inline-block w-20">{this.label()}</span>
+              {this.showConversionRate() && <span className="inline-block w-20">CR</span>}
+            </div>
           </div>
 
           <FlipMove className="flex-grow">