+++ test/plausible_web/controllers/api/stats_controller/browsers_test.exs
@@ -19,6 +19,22 @@ defmodule PlausibleWeb.Api.StatsController.BrowsersTest do
                %{"name" => "Firefox", "count" => 1, "percentage" => 33}
              ]
     end
+
+    test "calculates conversion_rate when filtering for goal", %{conn: conn, site: site} do
+      populate_stats(site, [
+        build(:pageview, user_id: 1, browser: "Chrome"),
+        build(:pageview, user_id: 2, browser: "Chrome"),
+        build(:event, user_id: 1, name: "Signup")
+      ])
+
+      filters = Jason.encode!(%{"goal" => "Signup"})
+
+      conn = get(conn, "/api/stats/#{site.domain}/browsers?period=day&filters=#{filters}")
+
+      assert json_response(conn, 200) == [
+               %{"name" => "Chrome", "count" => 1, "percentage" => 100, "conversion_rate" => 50.0}
+             ]
+    end
   end
 
   describe "GET /api/stats/:domain/browser-versions" do